requirements.txt
langchain
PyPDF2
python-dotenv
streamlit
openai
cassio 
datasets
tiktoken
langchain_community
langchain-openai